// context/types/post.ts
/* -----------------------------------------------------------------------------
XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX POSTS XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
----------------------------------------------------------------------------- */

/* POST: Project Types, Content & Content Provider Interface */
export type ITypes = {
    pages: string;
    posts: string;
    previewPage: string;
    previewPost: string;
};