// context/types/flexibleContentType.ts
/* -----------------------------------------------------------------------------
XXXXXXXXXXXXXXXXXXXXXXXXX FLEXIBLE CONTENT TYPES XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
----------------------------------------------------------------------------- */

/* FLEXIBLE CONTENT TYPES: Content & Content Provider Interface */
export type ITypes = {
    pages: string;
    previewPage: string;
    previewPost: string;
};